Code changes to check if a provided string is present in given 2D array.
Each char in the 2D array is shifted by 1 and then code checks if the
provided string is present in the given array.

This process is followed until word is found or the 2D array has
completed one full right shift whichever occurs first.

Return value: Returns x-cordinate, y-coordinate of the 1st character of
the string and number of right shift rotation required to match the
word.